chore: Update project configuration and documentation
This commit enhances the project setup by adding new scripts for local development in `package.json`, including commands for Vercel and Bun. It also updates the `.gitignore` file to exclude the `.vercel` directory. The README is improved with new sections for local and Docker development, providing clearer instructions for setting up the project. Additionally, the Vite configuration is modified to include a new API proxy, and the Infobase context is updated to use a local API during development. Various VSCode settings and extensions are also added for improved development experience.